upd: rename core.py to fix import error on taipy-enterprise
