+++ src/taipy/core/_core.py
